pyproject.toml
[project]
name = "gemini-mcp"
version = "1.1.1"
description = "MCP server enabling Claude-Gemini AI collaboration"
authors = [
    {name = "Abdulrahman Diaa", email = "a2diaa@icloud.com"},
]
dependencies = [
    "mcp[cli]>=0.9.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
gemini-mcp = "server:main"